refactor: Move `Token` to separate module.
It's still exported by the root module to prevent a breaking change.
